src/streaming/requirements.txt
avro-python3==1.10.2
requests==2.31.0
confluent-kafka==2.3.0
urllib3==2.1.0
gspread==5.12.2
google-auth==2.25.2
google-cloud-bigquery==3.14.1
oauth2client==4.1.3
